+++ relational-query/src/Database/Relational/Query.hs
@@ -25,6 +25,7 @@ module Database.Relational.Query (
   module Database.Relational.Query.Monad.Trans.Aggregating,
   module Database.Relational.Query.Monad.Trans.Ordering,
   module Database.Relational.Query.Monad.Trans.Assigning,
+  module Database.Relational.Query.Monad.BaseType,
   module Database.Relational.Query.Monad.Type,
   module Database.Relational.Query.Monad.Simple,
   module Database.Relational.Query.Monad.Aggregate,
@@ -57,16 +58,21 @@ import Database.Relational.Query.Projection (Projection, list)
 import Database.Relational.Query.Projectable
 import Database.Relational.Query.ProjectableExtended
 import Database.Relational.Query.Monad.Class
-  (MonadQualify, MonadRestrict, MonadQuery, MonadAggregate,
-   distinct, all', on, wheres, groupBy, having, restrict, onE, wheresE, havingE)
+  (MonadQualify,
+   MonadRestrict, wheres, having, restrict,
+   MonadAggregate, groupBy, groupBy',
+   MonadQuery, query', queryMaybe',
+   MonadPartition, partitionBy,
+   distinct, all', on, onE, wheresE, havingE)
 import Database.Relational.Query.Monad.Trans.Aggregating
-  (groupBy', key, key', set, bkey, rollup, cube, groupingSets)
+  (key, key', set, bkey, rollup, cube, groupingSets)
 import Database.Relational.Query.Monad.Trans.Ordering (orderBy, asc, desc)
 import Database.Relational.Query.Monad.Trans.Assigning (assignTo, (<-#))
+import Database.Relational.Query.Monad.BaseType
 import Database.Relational.Query.Monad.Type
 import Database.Relational.Query.Monad.Simple (QuerySimple, SimpleQuery)
 import Database.Relational.Query.Monad.Aggregate
-  (QueryAggregate, AggregatedQuery, Window, partitionBy, over)
+  (QueryAggregate, AggregatedQuery, Window, over)
 import Database.Relational.Query.Monad.Unique (QueryUnique)
 import Database.Relational.Query.Monad.Restrict (Restrict)
 import Database.Relational.Query.Monad.Assign (Assign)