+++ relational-query/src/Database/Relational/Query/Monad/Aggregate.hs
@@ -23,7 +23,7 @@ module Database.Relational.Query.Monad.Aggregate (
 
   toSubQuery,
 
-  Window, partitionBy, over
+  Window, over
   ) where
 
 import Data.Functor.Identity (Identity (runIdentity))
@@ -38,16 +38,15 @@ import Database.Relational.Query.Sub (SubQuery, aggregatedSubQuery, JoinProduct)
 import qualified Database.Relational.Query.Sub as SubQuery
 import Database.Relational.Query.Projectable (PlaceHolders, SqlProjectable)
 
-import Database.Relational.Query.Monad.Class (MonadRestrict(..), MonadQualify(..), MonadPartition (..))
-import Database.Relational.Query.Monad.Trans.Join (join')
+import Database.Relational.Query.Monad.Class (MonadRestrict(..))
 import Database.Relational.Query.Monad.Trans.Restricting
   (Restrictings, restrictings, extractRestrict)
 import Database.Relational.Query.Monad.Trans.Aggregating
-  (aggregatings, extractAggregateTerms, AggregatingSetT, PartitioningSet)
+  (extractAggregateTerms, AggregatingSetT, PartitioningSet)
 import Database.Relational.Query.Monad.Trans.Ordering
-  (Orderings, orderings, extractOrderingTerms)
-import Database.Relational.Query.Monad.Type
-  (ConfigureQuery, askConfig, QueryCore, extractCore, OrderedQuery)
+  (Orderings, extractOrderingTerms)
+import Database.Relational.Query.Monad.BaseType (ConfigureQuery, askConfig)
+import Database.Relational.Query.Monad.Type (QueryCore, extractCore, OrderedQuery)
 
 
 -- | Aggregated query monad type.
@@ -59,17 +58,9 @@ type AggregatedQuery p r = OrderedQuery Aggregated (Restrictings Aggregated (Agg
 -- | Partition monad type for partition-by clause.
 type Window           c = Orderings c (PartitioningSet c)
 
--- | Lift from qualified table forms into 'QueryAggregate'.
-aggregatedQuery :: ConfigureQuery a -> QueryAggregate a
-aggregatedQuery =  orderings . restrictings . aggregatings . restrictings . join'
-
 -- | Restricted 'MonadRestrict' instance.
 instance MonadRestrict Flat q => MonadRestrict Flat (Restrictings Aggregated q) where
   restrict = restrictings . restrict
-
--- | Instance to lift from qualified table forms into 'QueryAggregate'.
-instance MonadQualify ConfigureQuery QueryAggregate where
-  liftQualify = aggregatedQuery
 
 extract :: AggregatedQuery p r
         -> ConfigureQuery (((((((PlaceHolders p, Projection Aggregated r), OrderingTerms),
@@ -91,10 +82,6 @@ toSubQuery q = do
   (((((((_ph, pj), ot), grs), ag), rs), pd), da) <- extract q
   c <- askConfig
   return $ aggregatedSubQuery c (Projection.untype pj) da pd rs ag grs ot
-
--- | Add /PARTITION BY/ term into context.
-partitionBy :: Projection c r -> Window c ()
-partitionBy =  mapM_ unsafeAddPartitionKey . Projection.columns
 
 extractWindow :: Window c a -> ((a, OrderingTerms), [AggregateColumnRef])
 extractWindow =  runIdentity . extractAggregateTerms . extractOrderingTerms