+++ relational-query/src/Database/Relational/Query/Relation.hs
@@ -24,17 +24,12 @@ module Database.Relational.Query.Relation (
 
   uniqueRelation', aggregatedUnique,
 
-  dump,
-
-  sqlFromRelationWith, sqlFromRelation,
-
   -- * Query using relation
-  query, query', queryMaybe, queryMaybe', queryList, queryList', queryScalar, queryScalar',
+  query, queryMaybe, queryList, queryList', queryScalar, queryScalar',
   uniqueQuery', uniqueQueryMaybe',
 
   -- * Direct style join
   JoinRestriction,
-  rightPh, leftPh,
   inner', left', right', full',
   inner, left, right, full,
   on',
@@ -50,19 +45,20 @@ module Database.Relational.Query.Relation (
 import Control.Applicative ((<$>))
 
 import Database.Relational.Query.Context (Flat, Aggregated)
-import Database.Relational.Query.Monad.Type (ConfigureQuery, configureQuery, qualifyQuery)
+import Database.Relational.Query.Monad.BaseType
+  (ConfigureQuery, qualifyQuery,
+   Relation, unsafeTypeRelation, untypeRelation)
 import Database.Relational.Query.Monad.Class
-  (MonadQualify (liftQualify), MonadQualifyUnique (liftQualifyUnique), MonadQuery (unsafeSubQuery), on)
+  (MonadQualify (liftQualify), MonadQuery (query', queryMaybe'), on)
 import Database.Relational.Query.Monad.Simple (QuerySimple, SimpleQuery)
 import qualified Database.Relational.Query.Monad.Simple as Simple
 import Database.Relational.Query.Monad.Aggregate (QueryAggregate, AggregatedQuery)
 import qualified Database.Relational.Query.Monad.Aggregate as Aggregate
-import Database.Relational.Query.Monad.Unique (QueryUnique)
+import Database.Relational.Query.Monad.Unique (QueryUnique, unsafeUniqueSubQuery)
 import qualified Database.Relational.Query.Monad.Unique as Unique
 
-import Database.Relational.Query.Component (Config, defaultConfig, Duplication (Distinct, All))
+import Database.Relational.Query.Component (Duplication (Distinct, All))
 import Database.Relational.Query.Table (Table, TableDerivable, derivedTable)
-import Database.Relational.Query.Internal.SQL (StringSQL, showStringSQL)
 import Database.Relational.Query.Internal.Product (NodeAttr(Just', Maybe))
 import Database.Relational.Query.Sub (SubQuery)
 import qualified Database.Relational.Query.Sub as SubQuery
@@ -77,13 +73,9 @@ import Database.Relational.Query.Projectable
 import Database.Relational.Query.ProjectableExtended ((!))
 
 
--- | Relation type with place-holder parameter 'p' and query result type 'r'.
-newtype Relation p r = SubQuery (ConfigureQuery SubQuery)
-
-
 -- | Simple 'Relation' from 'Table'.
 table :: Table r -> Relation () r
-table =  SubQuery . return . SubQuery.fromTable
+table =  unsafeTypeRelation . return . SubQuery.fromTable
 
 -- | Infered 'Relation'.
 derivedRelation :: TableDerivable r => Relation () r
@@ -96,47 +88,11 @@ tableOf =  const derivedTable
 placeHoldersFromRelation :: Relation p r -> PlaceHolders p
 placeHoldersFromRelation =  const unsafePlaceHolders
 
--- | Sub-query Qualify monad from relation.
-subQueryQualifyFromRelation :: Relation p r -> ConfigureQuery SubQuery
-subQueryQualifyFromRelation =  d  where
-  d (SubQuery qsub)   = qsub
-
--- -- | Sub-query from relation.
--- subQueryFromRelation :: Relation p r -> SubQuery
--- subQueryFromRelation =  configureQuery . subQueryQualifyFromRelation
-
--- | Basic monadic join operation using 'MonadQuery'.
-queryWithAttr :: (MonadQualify ConfigureQuery m, MonadQuery m)
-              => NodeAttr
-              -> Relation p r
-              -> m (PlaceHolders p, Projection Flat r)
-queryWithAttr attr = unsafeAddPlaceHolders . run where
-  run rel = do
-    q <- liftQualify $ do
-      sq <- subQueryQualifyFromRelation rel
-      qualifyQuery sq
-    unsafeSubQuery attr q
-  -- d (Relation q) = unsafeMergeAnotherQuery attr q
-
--- | Join sub-query with place-holder parameter 'p'. query result is not 'Maybe'.
-query' :: (MonadQualify ConfigureQuery m, MonadQuery m)
-       => Relation p r
-       -> m (PlaceHolders p, Projection Flat r)
-query' =  queryWithAttr Just'
-
 -- | Join sub-query. Query result is not 'Maybe'.
 query :: (MonadQualify ConfigureQuery m, MonadQuery m)
       => Relation () r
       -> m (Projection Flat r)
 query =  fmap snd . query'
-
--- | Join sub-query with place-holder parameter 'p'. Query result is 'Maybe'.
-queryMaybe' :: (MonadQualify ConfigureQuery m, MonadQuery m)
-            => Relation p r
-            -> m (PlaceHolders p, Projection Flat (Maybe r))
-queryMaybe' pr =  do
-  (ph, pj) <- queryWithAttr Maybe pr
-  return (ph, Projection.just pj)
 
 -- | Join sub-query. Query result is 'Maybe'.
 --   The combinations of 'query' and 'queryMaybe' express
@@ -158,7 +114,7 @@ queryMaybe =  fmap snd . queryMaybe'
 queryList0 :: MonadQualify ConfigureQuery m => Relation p r -> m (ListProjection (Projection c) r)
 queryList0 =  liftQualify
               . fmap Projection.unsafeListFromSubQuery
-              . subQueryQualifyFromRelation
+              . untypeRelation
 
 -- | List sub-query, for /IN/ and /EXIST/ with place-holder parameter 'p'.
 queryList' :: MonadQualify ConfigureQuery m
@@ -179,7 +135,7 @@ addUnitPH =  ((,) unitPlaceHolder <$>)
 
 -- | Finalize 'QuerySimple' monad and generate 'Relation' with place-holder parameter 'p'.
 relation' :: SimpleQuery p r -> Relation p r
-relation' =  SubQuery . Simple.toSubQuery
+relation' =  unsafeTypeRelation . Simple.toSubQuery
 
 -- | Finalize 'QuerySimple' monad and generate 'Relation'.
 relation :: QuerySimple (Projection Flat r) -> Relation () r
@@ -187,7 +143,7 @@ relation =  relation' . addUnitPH
 
 -- | Finalize 'QueryAggregate' monad and geneate 'Relation' with place-holder parameter 'p'.
 aggregateRelation' :: AggregatedQuery p r -> Relation p r
-aggregateRelation' =  SubQuery . Aggregate.toSubQuery
+aggregateRelation' =  unsafeTypeRelation . Aggregate.toSubQuery
 
 -- | Finalize 'QueryAggregate' monad and geneate 'Relation'.
 aggregateRelation :: QueryAggregate (Projection Aggregated r) -> Relation () r
@@ -196,18 +152,6 @@ aggregateRelation =  aggregateRelation' . addUnitPH
 
 -- | Restriction function type for direct style join operator.
 type JoinRestriction a b = Projection Flat a -> Projection Flat b -> Projection Flat (Maybe Bool)
-
-unsafeCastPlaceHolder :: Relation a r -> Relation b r
-unsafeCastPlaceHolder =  d  where
-  d (SubQuery q)      = SubQuery q
-
--- | Simplify placeholder type applying left identity element.
-rightPh :: Relation ((), p) r -> Relation p r
-rightPh =  unsafeCastPlaceHolder
-
--- | Simplify placeholder type applying right identity element.
-leftPh :: Relation (p, ()) r -> Relation p r
-leftPh =  unsafeCastPlaceHolder
 
 -- | Basic direct join operation with place-holder parameters.
 join' :: (qa -> QuerySimple (PlaceHolders pa, Projection Flat a))
@@ -303,9 +247,9 @@ unsafeLiftAppend :: (SubQuery -> SubQuery -> SubQuery)
            -> Relation p a
            -> Relation q a
            -> Relation r a
-unsafeLiftAppend op a0 a1 = SubQuery $ do
-  s0 <- subQueryQualifyFromRelation a0
-  s1 <- subQueryQualifyFromRelation a1
+unsafeLiftAppend op a0 a1 = unsafeTypeRelation $ do
+  s0 <- untypeRelation a0
+  s1 <- untypeRelation a1
   return $ s0 `op` s1
 
 liftAppend :: (SubQuery -> SubQuery -> SubQuery)
@@ -373,29 +317,10 @@ infixl 8 `intersect`, `intersectAll`
 infixl 7 `union'`, `except'`, `unionAll'`, `exceptAll'`
 infixl 8 `intersect'`, `intersectAll'`
 
--- | Generate SQL string from 'Relation' with configuration.
-sqlFromRelationWith :: Relation p r -> Config -> StringSQL
-sqlFromRelationWith (SubQuery qsub) =  configureQuery $ SubQuery.showSQL <$> qsub
-
--- | SQL string from 'Relation'.
-sqlFromRelation :: Relation p r -> StringSQL
-sqlFromRelation =  (`sqlFromRelationWith` defaultConfig)
-
--- | Dump internal structure tree.
-dump :: Relation p r -> String
-dump =  show . (`configureQuery` defaultConfig) . subQueryQualifyFromRelation
-
-instance Show (Relation p r) where
-  show = showStringSQL . sqlFromRelation
-
 {-
 -- | Get projection width from 'Relation'.
 width :: Relation p r -> Int
 width =  SubQuery.width . subQueryFromRelation
-
--- | Finalize internal Query monad.
-nested :: Relation p r -> Relation p r
-nested =  SubQuery . subQueryFromRelation
 -}
 
 -- | Unique relation type to compose scalar queries.
@@ -410,34 +335,31 @@ unUnique :: UniqueRelation p c r -> Relation p r
 unUnique (Unique r) = r
 
 -- | Basic monadic join operation using 'MonadQuery'.
-uniqueQueryWithAttr :: MonadQualifyUnique ConfigureQuery m
-                    => NodeAttr
+uniqueQueryWithAttr :: NodeAttr
                     -> UniqueRelation p c r
-                    -> m (PlaceHolders p, Projection c r)
+                    -> QueryUnique (PlaceHolders p, Projection c r)
 uniqueQueryWithAttr attr = unsafeAddPlaceHolders . run where
   run rel = do
-    q <- liftQualifyUnique $ do
-      sq <- subQueryQualifyFromRelation (unUnique rel)
+    q <- liftQualify $ do
+      sq <- untypeRelation (unUnique rel)
       qualifyQuery sq
-    Projection.unsafeChangeContext <$> unsafeSubQuery attr q
+    Projection.unsafeChangeContext <$> unsafeUniqueSubQuery attr q
 
 -- | Join unique sub-query with place-holder parameter 'p'.
-uniqueQuery' :: MonadQualifyUnique ConfigureQuery m
-             => UniqueRelation p c r
-             -> m (PlaceHolders p, Projection c r)
+uniqueQuery' :: UniqueRelation p c r
+             -> QueryUnique (PlaceHolders p, Projection c r)
 uniqueQuery' = uniqueQueryWithAttr Just'
 
 -- | Join unique sub-query with place-holder parameter 'p'. Query result is 'Maybe'.
-uniqueQueryMaybe' :: MonadQualifyUnique ConfigureQuery m
-                  => UniqueRelation p c r
-                  -> m (PlaceHolders p, Projection c (Maybe r))
+uniqueQueryMaybe' :: UniqueRelation p c r
+                  -> QueryUnique (PlaceHolders p, Projection c (Maybe r))
 uniqueQueryMaybe' pr =  do
   (ph, pj) <- uniqueQueryWithAttr Maybe pr
   return (ph, Projection.just pj)
 
 -- | Finalize 'QueryUnique' monad and generate 'UniqueRelation'.
 uniqueRelation' :: QueryUnique (PlaceHolders p, Projection c r) -> UniqueRelation p c r
-uniqueRelation' =  unsafeUnique . SubQuery . Unique.toSubQuery
+uniqueRelation' =  unsafeUnique . unsafeTypeRelation . Unique.toSubQuery
 
 -- | Aggregated 'UniqueRelation'.
 aggregatedUnique :: Relation ph r
@@ -454,7 +376,7 @@ queryScalar' :: (MonadQualify ConfigureQuery m, ScalarDegree r)
              -> m (PlaceHolders p, Projection c (Maybe r))
 queryScalar' ur =
   unsafeAddPlaceHolders . liftQualify $
-  Projection.unsafeFromScalarSubQuery <$> subQueryQualifyFromRelation (unUnique ur)
+  Projection.unsafeFromScalarSubQuery <$> untypeRelation (unUnique ur)
 
 -- | Scalar sub-query.
 queryScalar :: (MonadQualify ConfigureQuery m, ScalarDegree r)