+++ relational-query/ChangeLog.md
@@ -1,3 +1,13 @@
+## 0.6.0.0
+
+- Increase type safety of interfaces.
+- Simplify interfaces arond unique query.
+
+## 0.5.2.0
+
+- Add MonadTrans instance of QueryJoin.
+- Update links about Opaleye.
+
 ## 0.5.1.1
 
 - Update documentation.